# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estate Hub — Real Estate ERP + CRM SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: getProperties
      tags: [properties]
      summary: List properties
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: property_type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyListResponse"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a single property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /ownerships:
    get:
      operationId: getOwnerships
      tags: [ownerships]
      summary: List property ownerships
      parameters:
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: property_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ownerships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ownership"
    post:
      operationId: createOwnership
      tags: [ownerships]
      summary: Create a property ownership record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnershipInput"
      responses:
        "201":
          description: Created ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownership"
  /ownerships/{id}:
    get:
      operationId: getOwnership
      tags: [ownerships]
      summary: Get a single ownership record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownership"
    patch:
      operationId: updateOwnership
      tags: [ownerships]
      summary: Update an ownership record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnershipUpdate"
      responses:
        "200":
          description: Updated ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ownership"

  # ─── PAYMENTS ───────────────────────────────────────────────
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List payment records
      parameters:
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: property_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a single payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─── INSTALLMENTS ────────────────────────────────────────────
...
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        project_name:
          type: string
        property_name:
          type: string
        unit_no:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        property_type:
          type: string
        price:
          type: number
        status:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - project_name
        - property_name
        - property_type
        - price
        - status
        - created_at
        - updated_at
    PropertyInput:
      type: object
      properties:
        project_name:
          type: string
        property_name:
          type: string
        unit_no:
          type: string
        location:
          type: string
        property_type:
          type: string
        price:
          type: number
        status:
          type: string
      required:
        - project_name
        - property_name
        - property_type
        - price
    PropertyUpdate:
      type: object
      properties:
        project_name:
          type: string
        property_name:
          type: string
        unit_no:
          type: string
        location:
          type: string
        property_type:
          type: string
        price:
          type: number
        status:
          type: string
...
    PropertyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ─── OWNERSHIP SCHEMAS ───────────────────────────────────
    Ownership:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        customer_id:
          type: integer
        property_id:
          type: integer
        customer_name:
          type: ["string", "null"]
        property_name:
          type: ["string", "null"]
        project_name:
          type: ["string", "null"]
        purchase_price:
          type: number
        purchase_date:
          type: string
        ownership_type:
          type: string
        created_at:
          type: string
      required:
        - id
        - company_id
        - customer_id
        - property_id
        - purchase_price
        - purchase_date
        - ownership_type
        - created_at
    OwnershipInput:
      type: object
      properties:
        customer_id:
          type: integer
        property_id:
          type: integer
        purchase_price:
          type: number
        purchase_date:
          type: string
        ownership_type:
          type: string
      required:
        - customer_id
        - property_id
        - purchase_price
        - purchase_date
        - ownership_type
    OwnershipUpdate:
      type: object
      properties:
        purchase_price:
          type: number
        purchase_date:
          type: string
        ownership_type:
          type: string

    # ─── PAYMENT SCHEMAS ─────────────────────────────────────
    Payment:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        customer_id:
          type: integer
        property_id:
          type: integer
        customer_name:
          type: ["string", "null"]
        property_name:
          type: ["string", "null"]
        total_amount:
          type: number
        paid_amount:
          type: number
        due_amount:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - customer_id
        - property_id
        - total_amount
        - paid_amount
        - due_amount
        - created_at
        - updated_at
    PaymentInput:
      type: object
      properties:
        customer_id:
          type: integer
        property_id:
          type: integer
        total_amount:
          type: number
        paid_amount:
          type: number
      required:
        - customer_id
        - property_id
        - total_amount
        - paid_amount
    PaymentUpdate:
      type: object
      properties:
        total_amount:
          type: number
        paid_amount:
          type: number

    # ─── INSTALLMENT SCHEMAS ─────────────────────────────────
...