# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Estate Hub — Real Estate ERP + CRM SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  company_id: number;
  project_name: string;
  property_name: string;
  /** @nullable */
  unit_no?: string | null;
  /** @nullable */
  location?: string | null;
  property_type: string;
  price: number;
  status: string;
  created_at: string;
  updated_at: string;
}