# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Estate Hub — Real Estate ERP + CRM SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  project_name?: string;
  property_name?: string;
  unit_no?: string;
  location?: string;
  property_type?: string;
  price?: number;
  status?: string;
}